refactor: split app rendering by environment
